+++ packages/@lwc/wds-core/src/node/plugins/mock-esm/const.js
@@ -1,2 +1,3 @@
 export const MOCK_CONTROLLER_PREFIX = '/virtual/mockController:';
 export const MOCK_STUB_PREFIX = '/virtual/mockStub:';
+export const UNMOCKED_ANNOTATION = 'unmocked';
