+++ packages/@lwc/wds-core/src/node/plugins/mock-esm/index.js
@@ -1,12 +1,10 @@
 import { resolve as pathResolve, dirname } from 'node:path';
-import { stringify as qsStringify } from 'node:querystring';
 import { makeMockControllerHandler } from './controller.js';
+import { makeMockedModuleHandler } from './mock-resolved.js';
 import { makeMockStubHandler } from './mock-stub.js';
 import { MOCK_CONTROLLER_PREFIX, MOCK_STUB_PREFIX } from './const.js';
-import { hasDefault, withoutDefault } from './util.js';
 
 const MOCK_IMPORT_PATTERN = /mock(\{ *([a-zA-Z0-9_]+( *, *)?)+\ *}):(.+)/;
-const UNMOCKED_ANNOTATION = 'unmocked';
 
 function parseExports(curlyWrappedNames) {
   return curlyWrappedNames
@@ -47,70 +45,6 @@ function makeRecursiveResolve(allPlugins) {
       importExists,
     };
   };
-}
-
-const buildMockForResolved = (absPathToUnmockedOriginal, exportedNames, queryString) => `
-import * as __original__ from '${absPathToUnmockedOriginal}${queryString}';
-
-${withoutDefault(exportedNames)
-  .map((name) => `export let ${name} = __original__['${name}'];`)
-  .join('\n')}
-${
-  hasDefault(exportedNames)
-    ? `
-let __liveDefault__ = __original__.default;
-const __hasDefault__ = true;
-export { __liveDefault__ as default };
-`
-    : `
-const __hasDefault__ = false;
-`
-}
-
-export const __mock__ = {
-  __setters__: {
-${withoutDefault(exportedNames)
-  .map((name) => `    ${name}: (val) => { ${name} = val; },`)
-  .join('\n')}
-  },
-  set(key, val) {
-    if (key === 'default') {
-      if (__hasDefault__) {
-        __liveDefault__ = val;
-      }
-    } else {
-      __mock__.__setters__[key](val);
-    }
-  },
-  reset(key) {
-    if (key === 'default') {
-      if (__hasDefault__) {
-        __liveDefault__ = __original__.default;
-      }
-    } else {
-      __mock__.__setters__[key](__original__[key]);
-    }
-  },
-  resetAll() {
-    Object.keys(__mock__.__setters__).forEach(name => {
-      __mock__.reset(name);
-    });
-  },
-  async useImport(importUrl) {
-    const newExports = await import(importUrl);
-    Object.keys(__mock__.__setters__).forEach(name => {
-      __mock__.__setters__[name](newExports[name]);
-    });
-  },
-};
-`;
-
-function getUnmockedUri(absoluteUrl, queryParams) {
-  const newParams = {
-    ...queryParams,
-    [UNMOCKED_ANNOTATION]: '1',
-  };
-  return `${absoluteUrl}?${qsStringify(newParams)}`;
 }
 
 export default ({ rootDir }) => {
@@ -162,23 +96,7 @@ export default ({ rootDir }) => {
 
   const serveMockController = makeMockControllerHandler({ mockedModules, rootDir });
   const serveMockStub = makeMockStubHandler({ mockedModules });
-
-  const serveMockedModule = (pathname, queryParams) => {
-    if (!mockedModules.has(pathname)) {
-      return;
-    }
-    // If the URL is annotated with `unmocked=1`, this passes through to the original
-    // source code. This allows the mocked version to access the stub or the actual,
-    // unmocked exports.
-    if (queryParams[UNMOCKED_ANNOTATION]) {
-      return;
-    }
-    const { exportedNames } = mockedModules.get(pathname);
-
-    const queryString = Object.keys(queryParams).length ? `?${qsStringify(queryParams)}` : '';
-
-    return buildMockForResolved(getUnmockedUri(pathname, queryParams), exportedNames, queryString);
-  };
+  const serveMockedModule = makeMockedModuleHandler({ mockedModules });
 
   return {
     name: 'mock-esm',
